refactor(hosts): refactors host similarly to home
Most of the logic for common modules in both NixOS and nix-darwin
(especially nix-darwin) were located in their library functions, which
shouldn't be the case. It's complicated to maintain and furthermore
confusing to understand why default behavior is located in a library.
Now the implementation simply imports modules from the directory and is
modularized into multiple modules.